Scale processes that use more than one cpu properly
Or maybe weirdly, but at least the line is visible.  If more than one
cpu is used than this processes' graph is scaled to show usage across
all cores, which means they are not comparable anymore with the other
graphs.  They might need a different colour or other highlighting
mechanism.