Print imprecise info about network usage (#4)
I could not figure out if this info is actual size on the wire in bytes,
if it is transmit or receive, if it is both (I _think_ but not sure),
and in general this is really only approximate info, but at least better
than before and now process level.  Hm.