Open menu with context actions on process label right-click
Figuring how to open a menu here on right click was a massive drag and
took all of today because it was not obvious to me at all that a gtk
label does not usually receive events.

So I needed a gtk event box and 🎉 it works!

The docs are here: https://lazka.github.io/pgi-docs/#Gtk-3.0/classes/Label.html#Gtk.Label.get_layout_offsets

> Of course you will need to create a Gtk.EventBox to receive the
> events, and pack the label inside it, since labels are windowless (they
> return False from Gtk.Widget.get_has_window()).

Of course...